+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/FileCreators/CO2FootprintSummary.groovy
@@ -2,6 +2,7 @@ package nextflow.co2footprint.FileCreators
 
 import groovy.util.logging.Slf4j
 import groovyx.gpars.agent.Agent
+import nextflow.co2footprint.CO2FootprintComputer
 import nextflow.co2footprint.Records.CO2EquivalencesRecord
 import nextflow.co2footprint.CO2FootprintConfig
 import nextflow.co2footprint.utils.Converter
@@ -39,15 +40,16 @@ class CO2FootprintSummary extends CO2FootprintFile {
     /**
      * Write the summary file with totals and options.
      *
-     * @param total_energy  Total expended energy during the run (Wh)
-     * @param total_co2     Total emitted CO₂e during the run (g)
-     * @param equivalences  Equivalences to CO₂ emissions
-     * @param config        CO2FootprintConfiguration
-     * @param version       Plugin version
+     * @param totalStats             Map containing total energy ('energy') in Wh and total CO₂ emissions ('co2e') in grams.
+     * @param co2FootprintComputer   CO2FootprintComputer instance for calculating equivalences.
+     * @param config                 CO2FootprintConfig instance with plugin configuration.
+     * @param version                Plugin version string.
      */
-    void write(Map<String, Double> totalStats, CO2EquivalencesRecord equivalences, CO2FootprintConfig config, String version) {
+    void write(Map<String, Double> totalStats, CO2FootprintComputer co2FootprintComputer, CO2FootprintConfig config, String version) {
         // Launch the agent (for thread safety, though only one write is performed)
         summaryWriter = new Agent<PrintWriter>(co2eSummaryFile)
+
+        CO2EquivalencesRecord equivalences = co2FootprintComputer.computeCO2footprintEquivalences(totalStats['co2e'])
 
         String outText = """\
         Total CO2e footprint measures of this workflow run (including cached tasks):